requirements.txt
# Core dependencies
numpy>=1.24.0
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
scikit-learn>=1.3.0
scikit-image>=0.21.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
Pillow>=10.0.0

# Deep Learning (optional, for CNN-based classification)
torch>=2.0.0
torchvision>=0.15.0

# Configuration & Utilities
PyYAML>=6.0
tqdm>=4.65.0
tabulate>=0.9.0

# Report generation
fpdf2>=2.7.0
